[ADH-4648] Implemented postgres partitioning for file access counts

+++ smart-metastore/src/main/java/org/smartdata/metastore/dao/CacheFileDao.java
@@ -41,7 +41,7 @@ public interface CacheFileDao
 
   void insert(List<CachedFileStatus> cachedFileStatusList);
 
-  int update(Long fid, Long lastAccessTime, Integer numAccessed);
+  int update(Long fid, Long lastAccessTime, long numAccessed);
 
   void update(Collection<AggregatedAccessCounts> events);
